requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy-financial==1.0.0
pandas==2.1.0
pydantic==2.3.0
python-dotenv==1.0.0
gunicorn==21.2.0